Move stratigraphy.ts to client logic
