+++ src/client/src/pages/detail/form/stratigraphy/components/lithologyTable/lithologyTableUtils.ts
@@ -1,8 +1,8 @@
 import { v4 as uuidv4 } from "uuid";
-import { BaseLayer, DepthLayer } from "../../../../../../api/stratigraphy.ts";
 import { FaciesDescription } from "../../faciesDescription.ts";
 import { LithologicalDescription } from "../../lithologicalDescription.ts";
 import { Lithology } from "../../lithology.ts";
+import { BaseLayer, DepthLayer } from "../../stratigraphy.ts";
 
 // TODO: What does this mean when building the depth layers?
 // Sort by fromDepth, tie-breaking on toDepth, then clamp toDepth to the next item's fromDepth wherever they overlap.
