+++ src/client/src/pages/detail/form/stratigraphy/components/lithologyTable/useLithologyTableState.ts
@@ -1,9 +1,9 @@
 import { useMemo, useRef, useState } from "react";
 import { v4 as uuidv4 } from "uuid";
-import { BaseLayer, DepthLayer } from "../../../../../../api/stratigraphy.ts";
 import { FaciesDescription } from "../../faciesDescription.ts";
 import { LithologicalDescription } from "../../lithologicalDescription.ts";
 import { Lithology } from "../../lithology.ts";
+import { BaseLayer, DepthLayer } from "../../stratigraphy.ts";
 import {
   createEmptyLithology,
   flagErrors,
