+++ src/client/src/pages/detail/form/stratigraphy/extraction/useBulkAddMutations.ts
@@ -1,10 +1,10 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { ApiError } from "../../../../../api/apiInterfaces.ts";
 import { fetchApiV2WithApiError } from "../../../../../api/fetchApiV2.ts";
-import { stratigraphiesQueryKey, Stratigraphy } from "../../../../../api/stratigraphy.ts";
 import { useResetTabStatus } from "../../../../../hooks/useResetTabStatus.ts";
 import { LithologicalDescription } from "../lithologicalDescription.ts";
 import { Lithology } from "../lithology.ts";
+import { stratigraphiesQueryKey, Stratigraphy } from "../stratigraphy.ts";
 
 interface StratigraphyInput {
   name: string;
