+++ src/client/src/pages/detail/form/stratigraphy/faciesDescription.ts
@@ -1,8 +1,8 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { fetchApiV2WithApiError } from "../../../../api/fetchApiV2.ts";
-import { BaseLayer } from "../../../../api/stratigraphy.ts";
 import { Codelist } from "../../../../components/codelist.ts";
 import { useResetTabStatus } from "../../../../hooks/useResetTabStatus.ts";
+import { BaseLayer } from "./stratigraphy.ts";
 
 export interface FaciesDescription extends BaseLayer {
   description?: string;
