+++ src/client/src/pages/detail/form/stratigraphy/lithostratigraphy/lithostratigraphyPanel.jsx
@@ -1,10 +1,10 @@
 import { useTranslation } from "react-i18next";
 import { Typography } from "@mui/material";
-import { useLithostratigraphies } from "../../../../../api/stratigraphy.ts";
 import NavigationChild from "../navigationChild.jsx";
 import NavigationContainer from "../navigationContainer.jsx";
 import NavigationLens from "../navigationLens.jsx";
 import Scale from "../scale.jsx";
+import { useLithostratigraphies } from "../stratigraphy.ts";
 import LithostratigraphyEditProfile from "./lithostratigraphyEditProfile.jsx";
 import LithostratigraphyViewProfile from "./lithostratigraphyViewProfile.jsx";
