+++ src/client/src/pages/detail/form/stratigraphy/lithostratigraphy/lithostratigraphyViewProfile.jsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { Box, LinearProgress } from "@mui/material";
-import { useLithostratigraphies } from "../../../../../api/stratigraphy.ts";
+import { useLithostratigraphies } from "../stratigraphy.ts";
 
 const handleColor = layer =>
   `rgb(${(JSON.parse(layer?.lithostratigraphy?.conf ?? null)?.color ?? [255, 255, 255]).join()})`;
