+++ src/client/src/pages/detail/form/stratigraphy/stratigraphyUtils.ts
@@ -1,6 +1,6 @@
-import { BaseLayer } from "../../../../api/stratigraphy.ts";
 import { defaultRowHeight } from "./components/stratigraphyTablePrimitives.tsx";
 import { LayerDepth, Lithology } from "./lithology.ts";
+import { BaseLayer } from "./stratigraphy.ts";
 
 export const computeCellHeight = (fromDepth: number, toDepth: number, depths: LayerDepth[]) => {
   const startIndex = depths.findIndex(l => l.fromDepth === fromDepth);
